requirements.txt
aws-cdk-lib==2.160.0
constructs>=10.0.0,<11.0.0
boto3==1.18.0
moto==5.0.16
pytest==8.3.3